Split files so actor is separate
Beside the actor being separated, the Queryable
Message model is moved to `crate::store` because it's
related to the database.

Now it should be possible to have all db-related code
in `store`, qt/qml-model-related code in `model`, and
the actors and their events and such in `actor`.

The purpose of this exercise is an attempt to keep the
files smaller as new models get implemented.