Fix configuration to build WF in the sfdk build env
This branch and commit is meant as a reference to lookup what may be
missing to get WF build in sfdk. I don't think that it currently makes
sense to switch the build systems. At least the provided way would need
a lot of tweeking to get it production ready.