Fix saving profile key for e.g. contact being added by full group update
This was hit e.g. when a group was being synced to Whisperfish for the
first time, and the full group sync tries to add members, which tries
to create missing recipients. This tried to then add the profile keys
to the just-added recipients, which failed with "UUID or profile key
of GroupV2 member is not set".