fix(ci): Don't cache and restore target contents for build-sailfishos
At the current state of CI, every package is built, not fresh-ed.
This means that sccache provides the build result, not the target.
This means that there's no point caching and restoring target contents.
This means CI doesn't have to pass `target/` needlessly around.
There were no hits for the non-star-paths so remove those as well.